requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
openai>=1.0.0
groq>=0.5.0
anthropic>=0.20.0
requests>=2.31.0
beautifulsoup4>=4.12.0
razorpay>=1.3.0
